+++ feature_structure/__brick__/{{feature_name}}/data/datasources/remote/{{feature_name}}_remote_datasource.dart
@@ -1,23 +1,21 @@
-import 'package:{{project_name}}/core/core.dart';
+import 'package:dartz/dartz.dart';
+import 'package:{{project_name}}/core/errors/failures.dart';
 
-
-
-abstract interface class {{ feature_name.pascalCase() }}RemoteDataSource {
-
+abstract interface class {{feature_name.pascalCase()}}RemoteDataSource {
   Future<Either<Failure, void>> fetchData();
-
 }
 
+class {{feature_name.pascalCase()}}RemoteDataSourceImpl
+    implements {{feature_name.pascalCase()}}RemoteDataSource {
+  const {{feature_name.pascalCase()}}RemoteDataSourceImpl();
 
-class {{ feature_name.pascalCase() }}RemoteDataSourceImpl implements {{ feature_name.pascalCase() }}RemoteDataSource {
-  
-    @override
-    Future<Either<Failure, void>> fetchData() async {
-      try {
-        await Future.delayed(const Duration(seconds: 1));
-        return const Right(null);
-      } catch (e) {
-        return const Left<Failure, void>(ExampleFailure(message:"Failed to fetch data"));
-      }
+  @override
+  Future<Either<Failure, void>> fetchData() async {
+    try {
+      await Future<void>.delayed(const Duration(seconds: 1));
+      return const Right<Failure, void>(null);
+    } catch (e) {
+      return Left<Failure, void>(ServerFailure(message: 'Failed to fetch data: $e'));
     }
   }
+}