+++ feature_structure/__brick__/{{feature_name}}/domain/repositories/{{feature_name}}_repository.dart
@@ -1,8 +1,6 @@
-import 'package:{{project_name}}/core/core.dart';
+import 'package:dartz/dartz.dart';
+import 'package:{{project_name}}/core/errors/failures.dart';
 
-abstract interface class {{ feature_name.pascalCase() }}Repository {
-  
-
+abstract interface class {{feature_name.pascalCase()}}Repository {
   Future<Either<Failure, void>> fetchData();
-
 }