+++ feature_structure/__brick__/{{feature_name}}/domain/use_cases/{{feature_name.snakeCase()}}_use_cases.dart
@@ -1,17 +1,14 @@
+import 'package:dartz/dartz.dart';
+import 'package:{{project_name}}/core/errors/failures.dart';
+import '../repositories/{{feature_name.snakeCase()}}_repository.dart';
 
-import 'package:{{project_name}}/features/{{feature_name.snakeCase()}}/domain/repositories/{{feature_name.snakeCase()}}_repository.dart';
-import 'package:{{project_name}}/core/core.dart';
+class {{feature_name.pascalCase()}}UseCases {
+  const {{feature_name.pascalCase()}}UseCases({
+    required {{feature_name.pascalCase()}}Repository repository,
+  }) : _{{feature_name.camelCase()}}Repository = repository;
 
+  final {{feature_name.pascalCase()}}Repository _{{feature_name.camelCase()}}Repository;
 
-class {{ feature_name.pascalCase() }}UseCases {
-  
-  const {{ feature_name.pascalCase() }}UseCases({
-  required {{ feature_name.pascalCase() }}Repository repository,
-  }) : _{{ feature_name.camelCase() }}Repository = repository;
-
-  final {{ feature_name.pascalCase() }}Repository _{{ feature_name.camelCase() }}Repository;
-
-  Future<Either<Failure, void>> fetchData() async {
-    return await _{{ feature_name.camelCase() }}Repository.fetchData();
-  }
+  Future<Either<Failure, void>> fetchData() =>
+      _{{feature_name.camelCase()}}Repository.fetchData();
 }